demo/server/requirements.txt
# RTSM Demo Server Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pyzmq>=25.1.0
numpy>=1.24.0
opencv-python>=4.8.0
